Add additional endpoints for asynchronous processing
Users are starting to get to the point where they're hitting the 60
second idle timeout that seems to be cutting things off. Increasing the
timeout would be brittle, so this commit introduces a couple of
endpoints for asynchronously submitting 'jobs', which the frontend will
then poll the status of before reporting to the user.