Flatten request objects
Now that the frontend is going to have multiple sampling strategies, and
possibly other needs, I don't want things to be nested in weird ways.

Ideally I'd make a different request object for each type of request,
but my experience with Haskell isn't strong enough to make that a quick
undertaking.