Add logging, wait for Redis
I was able to get output in Travis CI showing that the container died:

```
$ docker run --rm -d -p 8080:8080 sweetpea/server
2019-02-21T06:00:17.140124992Z container create 6483f6bc984e8cbd48b9298417b53add4e4fde1cf25f0d70fdc9f4116dbee948 (image=sweetpea/server, name=sad_proskuriakova)
6483f6bc984e8cbd48b9298417b53add4e4fde1cf25f0d70fdc9f4116dbee948
2019-02-21T06:00:17.210404809Z network connect ff494b201a53db89ed17777baeb7221a22846ffce95b415ba1fa55088ef37002 (container=6483f6bc984e8cbd48b9298417b53add4e4fde1cf25f0d70fdc9f4116dbee948, name=bridge, type=bridge)
2019-02-21T06:00:17.539480379Z container start 6483f6bc984e8cbd48b9298417b53add4e4fde1cf25f0d70fdc9f4116dbee948 (image=sweetpea/server, name=sad_proskuriakova)
before_install.6
0.13s$ docker ps -a
2019-02-21T06:00:17.640573916Z container die 6483f6bc984e8cbd48b9298417b53add4e4fde1cf25f0d70fdc9f4116dbee948 (exitCode=1, image=sweetpea/server, name=sad_proskuriakova)
```

My guess is that redis hasn't finished starting by the time the server
tries to connect, so it fails to start.

I added a 5 second sleep to the server's startup to hopefully compensate
for this. It's a bit hokey, but the docker image doesn't have `nc` or
`sleep` on it, so this is the next best thing.

Also adds a helpful request logger to the server.