Use connect instead of checkedConnect
`checkedConnect` throws an error if the connection can't be
established, which is what is killing our CI builds. The sleep wasn't
working either, and it's brittle, unnecessary locally, etc.

Using `connect` will cause the connection to be established when first
needed. Furthermore, failures to connect cause a 500 response, but
don't crash the server. If the redis serer comes up in the meantime,
subsequent attempts will succeed.